Fix bad docker-cp
a84a6da had the side-effect of automatically creating directories up to
the mountpoint.  Without the volume `docker cp` will fail because the
directory doesn't exist.

We could create it and commit a new image, maybe even derive the base
image using a Dockerfile, or we could do a little dance with `docker
exec`.  In the end, we don't need that directory, we can just put the
file in /tmp, which is what this patch does.