Bind AWS credentials for AWS push stage
Looking at
https://jenkins.io/doc/pipeline/steps/credentials-binding/

(via
https://wiki.jenkins-ci.org/display/JENKINS/Credentials+Binding+Plugin)

It seems the $class argument is the name of the binding implementation
class.  All examples here ("UsernamePasswordMultiBinding",
"FileBinding") correspond to such classes:
https://github.com/jenkinsci/credentials-binding-plugin/tree/4424c20/src/main/java/org/jenkinsci/plugins/credentialsbinding/impl

Looking at the aws-credentials plugins, we see that it takes a
dependency on the credentials-binding plugin, which would imply that it
supports it and provides an implementation.
https://wiki.jenkins-ci.org/display/JENKINS/CloudBees+AWS+Credentials+Plugin

Indeed, here it is:
https://github.com/jenkinsci/aws-credentials-plugin/blob/be11c1b/src/main/java/com/cloudbees/jenkins/plugins/awscredentials/AmazonWebServicesCredentialsBinding.java

Now we have the $class name to use.

Reading the code, we also confirm that the default environment variable
names are already those we want (AWS_ACCESS_KEY_ID,
AWS_SECRET_ACCESS_KEY).  Nothing else to do, but if we wanted to change
the variable names we'd use the "accessKeyVariable" and
"secretKeyVariable" arguments.

This thing is poorly documented (well, not documented at all) but at
least the code is clean.