refactor(dockerfile): Don't set exposed ports
Already defined in the base image:
https://github.com/jodogne/OrthancDocker/blob/79b9ffd1fabd4af56134f53e33cc86b37f8e5129/orthanc-plugins/Dockerfile#L19

Note that those are only useful if we use the --publish-all/-P option in
docker-run(1) which allocates random ports on the host and maps them to
the exposed ports in the container.  No need to worry about that until
we need it (so it doesn't matter if upstream changes in this regard).

More info:  We should only add this if we can ensure those ports are
indeed bound to a listening socket, which we can only do if we override
the default config of the base (which we don't).